fix(pkg/kernelrelease): get unique list besides different source package url
